src/third_party/blink/web_tests/external/wpt/service-workers/service-worker/unregister-then-register-new-script.https-expected.txt

This is a testharness.js-based test.
PASS Registering a new script URL while an unregistered registration is in use
PASS Registering a new script URL that 404s does not resurrect unregistered registration
FAIL Registering a new script URL that fails to install does not resurrect unregistered registration assert_equals: registration.active expected null but got object "[object ServiceWorker]"
Harness: the test ran to completion.
